+++ app/assets/javascripts/mapserver/App/Search/Advanced.js
@@ -79,10 +79,10 @@ App.Search.Advanced = App.newClass({
 			  query_parts.push("ST_DistanceSphere(shape_geom, ST_MakePoint(" + this.center.lng() + ", " + this.center.lat() + ")) <= " + distance);
 			  this.clearOverlays();
 			  var distanceOptions = {
-				  strokeColor: "#FF0000",
+				  strokeColor: "#fe5900",
 				  strokeOpacity: 0.8,
 				  strokeWeight: 2,
-				  fillColor: "#FF0000",
+				  fillColor: "#fe5900",
 				  fillOpacity: 0.15,
 				  map: map,
 				  center: this.center,