+++ app/assets/javascripts/mapserver/initMap.js
@@ -344,7 +344,7 @@ function initMap() {
     map.enableKeyDragZoom({
         key: 'ctrl',
         boxStyle: {
-            border: "4px solid #D93C00",
+            border: "4px solid #fe5900",
             backgroundColor: "transparent",
             opacity: 1.0
         },