+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - Ruby 3.1.7
+- rebrand to Český orienťák
 
 ## [2.4.0] - 2024-11-14
 - Ruby 3.1.2